Update V0.1.0 - Command and Connect update
* Rewards are much more rewarding, as they now work the same as the
Bypass and Available text boxes; allow you to create the reward without
typing anything more than the end result.

* CONNECTORS! Finally got around to uncommenting some code, and making
it do fancy things. Upon creating a focus with a prerequisite you will
create a connector (solid for single and AND, dashed for OR). Mutually
exlcusives are still in the work, will possibly come next update.

* Connectors move with both parent and child focuses when using
edit-in-view buttons (reason code was commented out)

* Slight CSS tweaks