V1.1 - The coder's update
After a lot of feedback from people who both love and hate this tool, it
was pretty obvious the current reward/available/bypass builder is very
broken...

The builder has not changed, but a trial option for the rewards has been
implemented, which will allow users to code as if it were the text
editor, with placeholder text showing available implementations, which
they can either auto-fill and edit (if needed), or ignore them, and type
it in full.

Changelog:
* New code-oriented reward textbox auto-filler
* Slight UI changes, mainly in terms of overlays (should look better on
all screen sizes now)
* JS files split, so that the export to file function doesn't break the
entire website on terrible browsers (i.e. very old browsers)